+++ internal/templates/blocks/broker_export_test.go
@@ -1,8 +1,0 @@
-package blocks
-
-import "github.com/nathanhollows/Rapua/v8/blocks"
-
-// Export internal functions for testing purposes.
-func GetBrokerInfoReceived(state blocks.PlayerState) string {
-	return getBrokerInfoReceived(state)
-}